get app working with wsgi + docker compose
