--- services/web/manage.py
@@ -1,10 +1,0 @@
-from flask.cli import FlaskGroup
-
-from app import app
-
-
-cli = FlaskGroup(app)
-
-
-if __name__ == "__main__":
-    cli()
