+++ services/web/project/__init__.py
@@ -1,0 +1,20 @@
+import os
+from flask import Flask
+from flask_sqlalchemy import SQLAlchemy
+
+db = SQLAlchemy()
+
+def create_app():
+    """Construct the core application"""
+    app = Flask(__name__)
+    app.config['SQLALCHEMY_DATABASE_URI'] = os.getenv("DATABASE_URL")
+    app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
+    # db = SQLAlchemy(app)
+    db.init_app(app)
+
+
+    with app.app_context():
+        from . import routes  # Import routes
+        db.create_all()  # Create sql tables for our data models
+
+        return app
